refactor(batch-details): move logic into existing cmd
