+++ src/utils/mod.rs
@@ -7,6 +7,7 @@ use zksync_ethers_rs::{
 };
 
 pub(crate) mod balance;
+pub(crate) mod chain;
 pub(crate) mod config;
 pub(crate) mod contract;
 pub(crate) mod contracts;
